# api-gateway/api/docs/swagger.yaml
definitions:
...
  booking.PaymentCreateReq:
    properties:
      amount:
        type: integer
      booking_id:
        type: string
      payment_method:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  booking.PaymentGetAllRes:
    properties:
      payments:
        items:
          $ref: '#/definitions/booking.PaymentRes'
        type: array
    type: object
  booking.PaymentGetByIdRes:
    properties:
      payment:
        $ref: '#/definitions/booking.PaymentRes'
    type: object
  booking.PaymentRes:
    properties:
      amount:
        type: integer
      booking_id:
        type: string
      id:
        type: string
      payment_method:
        type: string
      status:
        type: string
      user_id:
        type: string
    type: object
  booking.ProviderCreateReq:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      description:
        type: string
      location:
        type: string
      user_id:
        type: string
    type: object
  booking.ProviderGetAllRes:
    properties:
      providers:
        items:
          $ref: '#/definitions/booking.ProviderRes'
        type: array
    type: object
  booking.ProviderGetByIdRes:
    properties:
      provider:
        $ref: '#/definitions/booking.ProviderRes'
    type: object
  booking.ProviderRes:
    properties:
      availability:
        type: string
      average_rating:
        type: number
      company_name:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /payment:
    post:
      consumes:
      - application/json
      description: Create a new payment
      parameters:
      - description: Payment data
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/booking.PaymentCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Payment created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create payment
      tags:
      - payment
  /payment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment by ID
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment cancelled
          schema:
            type: string
        "400":
          description: Invalid payment ID
          schema:
            type: string
        "404":
          description: Payment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete payment
      tags:
      - payment
    get:
      consumes:
      - application/json
      description: Get a payment
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.PaymentGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get payment
      tags:
      - payment
    put:
      consumes:
      - application/json
      description: Update an existing payment
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Payment not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update payment
      tags:
      - payment
  /payment/all:
    get:
      consumes:
      - application/json
      description: Get all payment
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.PaymentGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all payment
      tags:
      - payment
  /provider:
    post:
      consumes:
      - application/json
      description: Create a new provider
      parameters:
      - description: Provider data
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/booking.ProviderCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Provider created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create provider
      tags:
      - provider
  /provider/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a provider by ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Provider deleted
          schema:
            type: string
        "400":
          description: Invalid provider ID
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete provider
      tags:
      - provider
    get:
      consumes:
      - application/json
      description: Get a provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get provider
      tags:
      - provider
    put:
      consumes:
      - application/json
      description: Update an existing provider
      parameters:
      - description: Provider data
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/booking.ProviderCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Provider updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Provider not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update provider
      tags:
      - provider
  /provider/all:
    get:
      consumes:
      - application/json
      description: Get all provider
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Rating
        in: query
        name: rating
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.ProviderGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all provider
      tags:
      - provider
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"